requirements.txt
torch
torchvision
flask
flask-cors
pillow
numpy
opencv-python
segmentation-models-pytorch
gunicorn